requirements.txt
huggingface-hub
python-dotenv